chore: rename site builder to build_site.py
Keep a thin main.py alias so the current deploy.yml still works
until the whitelist overlay lands. Pin mkdocs-material[imaging].
