+++ requirements.txt
@@ -2,7 +2,7 @@ black==25.1.0
 Requests==2.32.3
 sortedcontainers==2.4.0
 urllib3==1.26.18
-mkdocs-material==9.6.22
+mkdocs-material[imaging]==9.6.22
 mkdocs-glightbox==0.3.7
 jieba==0.42.1
 beautifulsoup4==4.13.3
